refactor: added a fix to bug render
Sometimes, when the app is first loading the table and sends a request to the server, the server would receive a pageSize of 1 and a page of 10. This was not ideal and caused the table to render only 1 resource on the first load.
Now, it checks if the arguments are 1 (for pageSize) and 10 (for page) and would send a predefined query to the database, returning the default amount of 10 resources.