feat: added new routes for new password
Added a new route in loginRouter that stores the user rut in a token, then generates a link (the link should be sent in a email to the user). For now, the backend console.logs the link to be accessed via the terminal.
Also added a route to handle the token when accessing to the newPassword component in the frontend. It executes when the component is first rendered: if the token is valid, the user can continue to the page. If it isn't, it is redirected to the login component.
Also added the route handler for when the user requests a password change. The route receives the new password, hashes it, then searchs the user in the database and stores the hushed password there.